Migrated Truffle unit test to Hardhat
This repo contains only the necessities to deploy a mock `ERC20`, `ERC721`, and `FantomAuction` contract on a local blockchain and test the `FantomAuction` contract with the supplied unit test. This `FantomAuction.sol` contract has been modified in a few ways to allow for the test environment to properly run, temporarily disabled functionality isn't needed to test a functioning `FantomAuction.sol` contract. A list of all changes to the contract for the purposes of testing are included in the README.MD